// FrontEnd_Angular/src/app/modules/admin/apps/contacts/contacts.types.ts
export interface Contact
{
    id: string;
    avatar?: string | null;
    background?: string | null;
    name: string;
    emails?: {
        email: string;
        label: string;
    }[];
    phoneNumbers?: {
        country: string;
        phoneNumber: string;
        label: string;
    }[];
    title?: string;
    company?: string;
    birthday?: string | null;
    address?: string | null;
    notes?: string | null;
    tags: string[];
}

export interface Country
{
    id?: string;
    iso?: string;
    name?: string;
    code?: string;
    flagImagePos?: string;
}

export interface Tag
{
    id?: string;
    title?: string;
}